requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiomcache==0.8.1
aiosignal==1.3.1
anyio==3.7.1
asgi-sessions==1.1.3
asgi-tools==0.75.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Authlib==1.2.1
bidict==0.22.1
black==23.7.0
bleach==6.0.0
build==1.0.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
configparser==6.0.0
cryptography==41.0.3
DateTime==5.2
decorouter==0.0.2
dnspython==2.4.2
docutils==0.20.1
extensions==0.4
fastapi==0.103.1
fastapi-socketio==0.0.10
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
htmx==0.0.0
http-router==4.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
is-disposable-email==1.0.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.2.0
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
motor==3.3.1
multidict==6.0.4
mypy-extensions==1.0.0
nicegui==1.3.13
orjson==3.9.6
outcome==1.2.0
packaging==23.1
pathspec==0.11.2
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.3.0
pscript==0.7.7
pybase64==1.3.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
pyjson==1.3.0
PyJWT==2.8.0
pymongo==4.5.0
PyNaCl==1.5.0
pypickle==1.1.0
pyproject_hooks==1.0.0
pytest==7.4.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.7.0
python-multipart==0.0.6
python-socketio==5.9.0
pytz==2023.3.post1
PyYAML==6.0.1
readme-renderer==41.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
ruff==0.0.287
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
starlette-authlib==0.1.30
starlette-decoRouter==1.0.0
starlette-session
Starlette-WTF==0.4.3
trio==0.22.2
twine==4.0.2
types-contextvars==2.4.7.3
types-dataclasses==0.6.6
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.22.0
uvloop==0.17.0
vbuild==0.8.2
watchfiles==0.20.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.7
WTForms==3.0.1
wtforms-validators==1.0.0
yarl==1.9.2
zipp==3.16.2
zope.interface==6.0
